+++ gravitee-am-gateway/gravitee-am-gateway-handler/gravitee-am-gateway-handler-core/src/main/java/io/gravitee/am/gateway/handler/root/resources/endpoint/mfa/MFAChallengePostEndpoint.java
@@ -16,10 +16,8 @@
 package io.gravitee.am.gateway.handler.root.resources.endpoint.mfa;
 
 import io.gravitee.am.common.audit.EventType;
-import io.gravitee.am.common.factor.FactorDataKeys;
 import io.gravitee.am.common.jwt.JWT;
 import io.gravitee.am.common.utils.ConstantKeys;
-import io.gravitee.am.common.utils.MovingFactorUtils;
 import io.gravitee.am.factor.api.FactorContext;
 import io.gravitee.am.factor.api.FactorProvider;
 import io.gravitee.am.gateway.handler.common.email.EmailService;
@@ -28,39 +26,29 @@ import io.gravitee.am.gateway.handler.common.jwt.JWTService;
 import io.gravitee.am.gateway.handler.common.service.CredentialGatewayService;
 import io.gravitee.am.gateway.handler.common.service.DeviceGatewayService;
 import io.gravitee.am.gateway.handler.common.service.mfa.VerifyAttemptService;
-import io.gravitee.am.gateway.handler.common.utils.HashUtil;
 import io.gravitee.am.gateway.handler.common.vertx.core.http.VertxHttpServerRequest;
-import io.gravitee.am.gateway.handler.common.vertx.utils.UriBuilderRequest;
 import io.gravitee.am.gateway.handler.manager.deviceidentifiers.DeviceIdentifierManager;
 import io.gravitee.am.gateway.handler.root.service.user.UserService;
 import io.gravitee.am.identityprovider.api.DefaultUser;
-import io.gravitee.am.model.ApplicationFactorSettings;
 import io.gravitee.am.model.Credential;
 import io.gravitee.am.model.Factor;
 import io.gravitee.am.model.MFASettings;
-import io.gravitee.am.model.Reference;
 import io.gravitee.am.model.RememberDeviceSettings;
 import io.gravitee.am.model.Template;
 import io.gravitee.am.model.User;
 import io.gravitee.am.model.VerifyAttempt;
 import io.gravitee.am.model.factor.EnrolledFactor;
-import io.gravitee.am.model.factor.EnrolledFactorChannel;
-import io.gravitee.am.model.factor.EnrolledFactorChannel.Type;
 import io.gravitee.am.model.factor.EnrolledFactorSecurity;
 import io.gravitee.am.model.factor.FactorStatus;
 import io.gravitee.am.model.oidc.Client;
 import io.gravitee.am.service.AuditService;
 import io.gravitee.am.service.DomainDataPlane;
-import io.gravitee.am.service.exception.FactorNotFoundException;
 import io.gravitee.am.service.exception.MFAValidationAttemptException;
 import io.gravitee.am.service.reporter.builder.AuditBuilder;
 import io.gravitee.am.service.reporter.builder.MFAAuditBuilder;
 import io.gravitee.am.service.reporter.builder.gateway.VerifyAttemptAuditBuilder;
 import io.gravitee.am.service.utils.vertx.RequestUtils;
-import io.gravitee.common.http.HttpHeaders;
-import io.gravitee.common.util.Maps;
 import io.gravitee.gateway.api.el.EvaluableRequest;
-import io.netty.handler.codec.http.QueryStringDecoder;
 import io.reactivex.rxjava3.core.Maybe;
 import io.reactivex.rxjava3.core.Single;
 import io.vertx.core.AsyncResult;
@@ -70,62 +58,47 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.rxjava3.core.MultiMap;
 import io.vertx.rxjava3.core.http.Cookie;
 import io.vertx.rxjava3.core.http.HttpServerRequest;
-import io.vertx.rxjava3.core.http.HttpServerResponse;
 import io.vertx.rxjava3.ext.web.RoutingContext;
 import io.vertx.rxjava3.ext.web.common.template.TemplateEngine;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationContext;
 
-import java.util.Comparator;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 import static com.google.common.base.Strings.isNullOrEmpty;
 import static io.gravitee.am.common.audit.EventType.MFA_CHALLENGE;
 import static io.gravitee.am.common.audit.EventType.MFA_ENROLLMENT;
 import static io.gravitee.am.common.audit.EventType.MFA_MAX_ATTEMPT_REACHED;
-import static io.gravitee.am.common.factor.FactorSecurityType.RECOVERY_CODE;
-import static io.gravitee.am.common.factor.FactorSecurityType.SHARED_SECRET;
 import static io.gravitee.am.common.factor.FactorSecurityType.WEBAUTHN_CREDENTIAL;
 import static io.gravitee.am.common.factor.FactorType.FIDO2;
 import static io.gravitee.am.common.utils.ConstantKeys.DEVICE_ALREADY_EXISTS_KEY;
 import static io.gravitee.am.common.utils.ConstantKeys.DEVICE_ID;
 import static io.gravitee.am.common.utils.ConstantKeys.DEVICE_TYPE;
 import static io.gravitee.am.common.utils.ConstantKeys.ENROLLED_FACTOR_ID_KEY;
-import static io.gravitee.am.common.utils.ConstantKeys.ERROR_HASH;
 import static io.gravitee.am.common.utils.ConstantKeys.ERROR_PARAM_KEY;
 import static io.gravitee.am.common.utils.ConstantKeys.MFA_CHALLENGE_FAILED;
 import static io.gravitee.am.common.utils.ConstantKeys.PASSWORDLESS_CHALLENGE_KEY;
 import static io.gravitee.am.common.utils.ConstantKeys.PASSWORDLESS_CHALLENGE_USERNAME_KEY;
-import static io.gravitee.am.common.utils.ConstantKeys.TRANSACTION_ID_KEY;
 import static io.gravitee.am.common.utils.ConstantKeys.VERIFY_ATTEMPT_ERROR_PARAM_KEY;
 import static io.gravitee.am.common.utils.ConstantKeys.WEBAUTHN_CREDENTIAL_INTERNAL_ID_CONTEXT_KEY;
 import static io.gravitee.am.gateway.handler.common.utils.RoutingContextUtils.getEvaluableAttributes;
 import static io.gravitee.am.model.factor.FactorStatus.ACTIVATED;
 import static io.gravitee.am.model.factor.FactorStatus.PENDING_ACTIVATION;
-import static java.lang.Boolean.TRUE;
 import static java.util.Optional.ofNullable;
 
-public class MFAChallengePostEndpoint extends MFAEndpoint {
-
+public class MFAChallengePostEndpoint extends MFAChallengeEndpoint {
     private static final Logger logger = LoggerFactory.getLogger(MFAChallengePostEndpoint.class);
 
     private static final String REMEMBER_DEVICE_CONSENT = "rememberDeviceConsent";
     public static final String REMEMBER_DEVICE_CONSENT_ON = "on";
-    public static final String PREVIOUS_TRANSACTION_ID_KEY = "prev-tid";
 
-    private final FactorManager factorManager;
     private final UserService userService;
     private final ApplicationContext applicationContext;
     private final DeviceGatewayService deviceService;
-    private final DomainDataPlane domainDataPlane;
     private final CredentialGatewayService credentialService;
     private final VerifyAttemptService verifyAttemptService;
     private final EmailService emailService;
@@ -148,12 +121,10 @@ public class MFAChallengePostEndpoint extends MFAEndpoint {
                                     JWTService jwtService,
                                     String rememberDeviceCookieName
     ) {
-        super(engine);
+        super(factorManager, engine, domainDataPlane, auditService);
         this.applicationContext = applicationContext;
-        this.factorManager = factorManager;
         this.userService = userService;
         this.deviceService = deviceService;
-        this.domainDataPlane = domainDataPlane;
         this.credentialService = credentialService;
         this.verifyAttemptService = verifyAttemptService;
         this.emailService = emailService;
@@ -368,164 +339,6 @@ public class MFAChallengePostEndpoint extends MFAEndpoint {
                 );
     }
 
-
-    private Factor getFactor(RoutingContext routingContext, Client client, User endUser) {
-        // factor can be either in session (if user come from mfa/enroll or mfa/challenge/alternatives page)
-        // or from the user enrolled factor list
-        final String savedFactorId =
-                routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_ID_KEY) != null ?
-                        routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_ID_KEY) :
-                        routingContext.session().get(ConstantKeys.ALTERNATIVE_FACTOR_ID_KEY);
-
-        if (savedFactorId != null) {
-            return factorManager.getFactor(savedFactorId);
-        }
-
-        if (endUser.getFactors() == null) {
-            throw FactorNotFoundException.withMessage("No factor found for the end user");
-        }
-
-        // get the primary enrolled factor
-        // if there is no primary, select the first created
-        final var factors = hasApplicationFactorSettings(client) ? client.getFactorSettings()
-                .getApplicationFactors()
-                .stream()
-                .map(ApplicationFactorSettings::getId)
-                .collect(Collectors.toSet()) : Set.of();
-
-        final var enrolledFactors = endUser.getFactors()
-                .stream()
-                .filter(enrolledFactor -> factors.contains(enrolledFactor.getFactorId()))
-                .sorted(Comparator.comparing(EnrolledFactor::getCreatedAt))
-                .toList();
-
-        if (enrolledFactors.isEmpty()) {
-            throw FactorNotFoundException.withMessage("No factor found for the end user");
-        }
-
-        return enrolledFactors
-                .stream()
-                .filter(e -> TRUE.equals(e.isPrimary()))
-                .map(enrolledFactor -> factorManager.getFactor(enrolledFactor.getFactorId()))
-                .findFirst()
-                .orElse(factorManager.getFactor(enrolledFactors.get(0).getFactorId()));
-    }
-
-    private EnrolledFactor getEnrolledFactor(RoutingContext routingContext,
-                                             FactorProvider factorProvider,
-                                             Factor factor,
-                                             User endUser,
-                                             FactorContext factorContext,
-                                             boolean overrideMovingFactor) {
-        // enrolled factor can be either in session (if user come from mfa/enroll page)
-        // or from the user enrolled factor list
-        final String savedFactorId = routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_ID_KEY);
-        if (factor.getId().equals(savedFactorId)) {
-            EnrolledFactor enrolledFactor = new EnrolledFactor();
-            enrolledFactor.setFactorId(factor.getId());
-            enrolledFactor.setStatus(PENDING_ACTIVATION);
-            switch (factor.getFactorType()) {
-                case OTP:
-                    enrolledFactor.setSecurity(new EnrolledFactorSecurity(SHARED_SECRET,
-                            routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_SECURITY_VALUE_KEY)));
-                    break;
-                case SMS:
-                    enrolledFactor.setChannel(new EnrolledFactorChannel(Type.SMS,
-                            routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_PHONE_NUMBER)));
-                    break;
-                case CALL:
-                    enrolledFactor.setChannel(new EnrolledFactorChannel(Type.CALL,
-                            routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_PHONE_NUMBER)));
-                    break;
-                case EMAIL:
-                    enrolledFactor.setChannel(new EnrolledFactorChannel(Type.EMAIL,
-                            routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_EMAIL_ADDRESS)));
-                    break;
-                case RECOVERY_CODE:
-                    if (endUser.getFactors() != null) {
-                        Optional<EnrolledFactorSecurity> factorSecurity = endUser.getFactors()
-                                .stream()
-                                .filter(ftr -> ftr.getSecurity().getType().equals(RECOVERY_CODE))
-                                .map(EnrolledFactor::getSecurity)
-                                .findFirst();
-
-                        factorSecurity.ifPresent(enrolledFactor::setSecurity);
-                    }
-                    break;
-                default:
-            }
-
-            // if the factor provider uses a moving factor security mechanism,
-            // we ensure that every data has been shared with the user enrolled factor
-            if (factorProvider.useVariableFactorSecurity(factorContext)) {
-                String tid = routingContext.session().get(PREVIOUS_TRANSACTION_ID_KEY);
-                if (overrideMovingFactor) {
-                    tid = routingContext.get(TRANSACTION_ID_KEY);
-                    routingContext.session().put(PREVIOUS_TRANSACTION_ID_KEY, tid);
-                }
-                enrolledFactor.setSecurity(new EnrolledFactorSecurity(SHARED_SECRET,
-                        routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_SECURITY_VALUE_KEY)));
-                Map<String, Object> additionalData = new Maps.MapBuilder<String, Object>(new HashMap<>())
-                        .put(FactorDataKeys.KEY_MOVING_FACTOR, MovingFactorUtils.generateInitialMovingFactor(tid))
-                        .build();
-                getEnrolledFactor(factor, endUser).ifPresent(ef -> {
-                    additionalData.put(FactorDataKeys.KEY_EXPIRE_AT, ef.getSecurity().getData(FactorDataKeys.KEY_EXPIRE_AT, Long.class));
-                });
-                enrolledFactor.getSecurity().getAdditionalData().putAll(additionalData);
-            }
-
-            // if there is an extension phone number, add it to the enrolled factor
-            final String extensionPhoneNumber = routingContext.session().get(ConstantKeys.ENROLLED_FACTOR_EXTENSION_PHONE_NUMBER);
-            if (extensionPhoneNumber != null && enrolledFactor.getChannel() != null) {
-                var additionalData = new HashMap<String, Object>();
-                additionalData.put(ConstantKeys.MFA_ENROLLMENT_EXTENSION_PHONE_NUMBER, extensionPhoneNumber);
-                enrolledFactor.getChannel().setAdditionalData(additionalData);
-            }
-
-            enrolledFactor.setCreatedAt(new Date());
-            enrolledFactor.setUpdatedAt(enrolledFactor.getCreatedAt());
-            return enrolledFactor;
-        }
-
-        return getEnrolledFactor(factor, endUser)
-                .orElseThrow(() -> FactorNotFoundException.withMessage("No enrolled factor found for the end user"));
-    }
-
-    private Optional<EnrolledFactor> getEnrolledFactor(Factor factor, User endUser) {
-        if (endUser.getFactors() == null) {
-            return Optional.empty();
-        }
-
-        return endUser.getFactors()
-                .stream()
-                .filter(f -> factor.getId().equals(f.getFactorId()))
-                .findFirst();
-    }
-
-    @Override
-    public String getTemplateSuffix() {
-        return Template.MFA_CHALLENGE.template();
-    }
-
-    private void doRedirect(HttpServerResponse response, String url) {
-        response.putHeader(HttpHeaders.LOCATION, url).setStatusCode(302).end();
-    }
-
-    private void handleException(RoutingContext context, String errorKey, String errorValue) {
-        final HttpServerRequest req = context.request();
-        final HttpServerResponse resp = context.response();
-
-        // redirect to mfa challenge page with error message
-        QueryStringDecoder queryStringDecoder = new QueryStringDecoder(req.uri());
-        Map<String, String> parameters = new LinkedHashMap<>(queryStringDecoder.parameters().entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().get(0))));
-        parameters.put(errorKey, errorValue);
-        if (context.session() != null) {
-            context.session().put(ERROR_HASH, HashUtil.generateSHA256(errorValue));
-        }
-        String uri = UriBuilderRequest.resolveProxyRequest(req, req.path(), parameters, true);
-        doRedirect(resp, uri);
-    }
-
     private RememberDeviceSettings getRememberDeviceSettings(Client client) {
         return ofNullable(client.getMfaSettings())
                 .map(MFASettings::getRememberDevice)
@@ -656,24 +469,6 @@ public class MFAChallengePostEndpoint extends MFAEndpoint {
 
     private static boolean isEnrolling(EnrolledFactor enrolledFactor) {
         return enrolledFactor.getStatus() == PENDING_ACTIVATION;
-    }
-
-    private void updateAuditLog(RoutingContext routingContext, String type, User endUser, Client client, Factor factor, FactorContext factorContext, Throwable cause) {
-        final EnrolledFactor enrolledFactor = factorContext.getData(FactorContext.KEY_ENROLLED_FACTOR, EnrolledFactor.class);
-        final EnrolledFactorChannel channel = enrolledFactor.getChannel();
-
-        final MFAAuditBuilder builder = AuditBuilder.builder(MFAAuditBuilder.class)
-                .user(endUser)
-                .factor(factor)
-                .type(type)
-                .channel(channel)
-                .client(client)
-                .reference(Reference.domain(domainDataPlane.getDomain().getId()))
-                .ipAddress(routingContext)
-                .userAgent(routingContext)
-                .throwable(cause, channel);
-
-        auditService.report(builder);
     }
 
     private void auditRememberDevice(User endUser, Client client, Throwable cause) {
